Address review feedback from emcfarlane
- Tighten the requestBodyWriter field and newDuplexHTTPCall pipe-setup comments to the suggested wording.

- Move the concurrent Send/CloseWrite regression test into duplex_http_call_test.go, rewritten against the internal duplexHTTPCall API in package connect (dropping the dedicated duplex_http_call_race_test.go file). Verified that the test still reproduces the original nil-deref panic when the fix is reverted and passes cleanly (including under -race) with the fix in place.
